requirements.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.6.20
chardet==3.0.4
DateTime==4.3
defusedxml==0.6.0
dj-rest-auth==1.0.9
Django==3.0.7
django-admin-interface==0.12.2
django-allauth==0.42.0
django-colorfield==0.3.1
django-cors-headers==3.4.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.10
oauthlib==3.1.0
pycodestyle==2.6.0
python-decouple==3.3
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
stripe==2.49.0
toml==0.10.1
urllib3==1.25.9
zope.interface==5.1.0
